Allow dynamic outage messages from JSON
*This is a breaking change*

This change allows for customization of the maintenance message on a
host-by-host basis.  It does this by adding new values to the `inform`
json, and adding a new configuration value.

The new architecture assumes that there is a directory (communicated via
configuration) that contains the template files that will be served.
When it loads the `inform` json, it also expects that the json will
contain a `template` value that, when appended to the base configuration
directory, will give the location of a template file to display when the
service is in maintenance mode.

The json should also include a `header` and `message` value.  These
values will be used to control the message returned when maintenance
mode is enabled. The template file should also have two `%s` symbols
inside so that the header and message text can be included in the final
message.